Consolidate the [Explicit] sample into the [Skip] sample, because they are both examples of skip rules and because it is useful to demonstrate the way that multiple skip rules are prioritized.
