+++ src/Fixie.Samples/SamplesAssembly.cs
@@ -6,7 +6,6 @@
         {
             Apply<Categories.CustomConvention>();
             Apply<ColumnParameter.CustomConvention>();
-            Apply<Explicit.CustomConvention>();
             Apply<Inclusive.CustomConvention>();
             Apply<IoC.CustomConvention>();
             Apply<LowCeremony.CustomConvention>();
