+++ src/Fixie.Samples/Skipped/ExplicitAttribute.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Fixie.Samples.Explicit
+namespace Fixie.Samples.Skipped
 {
     [AttributeUsage(AttributeTargets.Method, Inherited = false)]
     public class ExplicitAttribute : Attribute { }
